requirements.txt
jira
pandas
requests
odfpy
lxml
trello
tabulate
markdown-it-py